Move `TagColor` data class to SubjectFormatter
